CF221 DXClient Update

+++ bin/dxclient.bat
@@ -1,30 +1,24 @@
 @echo off
 SETLOCAL EnableDelayedExpansion
-:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-:: Licensed to the Apache Software Foundation (ASF) under one  
-:: or more contributor license agreements.  See the NOTICE file
-:: distributed with this work for additional information       
-:: regarding copyright ownership.  The ASF licenses this file  
-:: to you under the Apache License, Version 2.0 (the           
-:: "License"); you may not use this file except in compliance  
-:: with the License.  You may obtain a copy of the License at  
-::                                                             
-::   http://www.apache.org/licenses/LICENSE-2.0                
-::                                                             
-:: Unless required by applicable law or agreed to in writing,  
-:: software distributed under the License is distributed on an 
-:: "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY      
-:: KIND, either express or implied.  See the License for the   
-:: specific language governing permissions and limitations     
-:: under the License.                                          
-:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-:: The script is a wrapper to run dxclient tool. 
+:: 
+::********************************************************************
+:: Licensed Materials - Property of HCL                              *
+::                                                                   *
+::  Copyright HCL Technologies Ltd. 2021, 2024. All Rights Reserved. *
+::                                                                   *
+::  Note to US Government Users Restricted Rights:                   *
+::                                                                   *
+::  Use, duplication or disclosure restricted by GSA ADP Schedule    *
+:: *******************************************************************
+::
+:: The script is a wrapper to run dxclient tool.
 ::
 
 SET IMAGE_NAME=hclcr.io/dx-public/dxclient
 SET IMAGE_TAG=<IMAGE_TAG>
 SET DXCLIENT=dxclient
 SET DATA_DIR=store
+SET MOUNTED_VOL=
 
 IF [!IMAGE_NAME!]==[] (
 	echo "Please set dxclient image name."
@@ -40,6 +34,7 @@ IF [!CONTAINER_RUNTIME!]==[] (
 IF [!VOLUME_DIR!]==[] (
 	SET VOLUME_DIR=store
 )
+
 :: allows calling program (i.e: gradle) to not use interactive param for non-TTY type env
 SET "INTERACTIVE=-it"
 IF ["!DXCLIENT_DISABLE_INTERACTIVE!"]==["true"] (
@@ -58,7 +53,7 @@ if NOT EXIST "%cd%"/!VOLUME_DIR! (
 )
 
 :: mount volume directory
-CALL SET volumeParams=-v "%cd%\!VOLUME_DIR!\:/!DXCLIENT!/!DATA_DIR!":Z
+CALL SET volumeParams=-v "%cd%\!VOLUME_DIR!\:/!DXCLIENT!/!VOLUME_DIR!":Z
 
 CALL SET "newCmd="
 
@@ -67,6 +62,43 @@ CALL SET environmentVars= -e TZ=!Timezone! -e VOLUME_DIR="!VOLUME_DIR!"
 
 :: Setting search value "\" for validating filepath
 CALL SET searchVal=\
+SET hostPath=%CD%
+set exportPathExists=0
+set count=0
+
+:loop
+if "%~1"=="" goto :done
+set /A count+=1
+if /I "%~1"=="-exportPath" (
+    set exportPathExists=1
+    if not "%~2"=="" (
+        set hostPath="%~2"
+        shift
+    )
+) else if /I "%~1"=="--exportPath" (
+    set exportPathExists=1
+    if not "%~2"=="" (
+        set hostPath="%~2"
+        shift
+    )
+) else if /I "%~1"=="-themePath" (
+    set exportPathExists=1
+    if not "%~2"=="" (
+        set hostPath="%~2"
+        shift
+    )
+) else if /I "%~1"=="--themePath" (
+    set exportPathExists=1
+    if not "%~2"=="" (
+        set hostPath="%~2"
+        shift
+    )
+)
+if "%~1"=="" goto :done
+set /A count+=1
+shift
+goto :loop
+:done
 
 :: Generate a separate volume point for each folder or file in the parameter
 :: Replace the folders in the parameter with the assigned container folder in volume parameter
@@ -75,8 +107,9 @@ CALL SET searchVal=\
 	IF EXIST %%i (
 		:: If input doesnot contain "\" it should not replace the host path with container path
         IF NOT "!input:%searchVal%=!"=="!input!" (
-            CALL SET "newCmd=!newCmd! ^"/!DXCLIENT!/!DATA_DIR!/%%~nxi^""
-            CALL SET volumeParams=!volumeParams! -v "%%~fi:/!DXCLIENT!/!DATA_DIR!/%%~nxi":Z
+            CALL SET "newCmd=!newCmd! ^"/!DXCLIENT!/!VOLUME_DIR!/%%~nxi^""
+            CALL SET volumeParams=!volumeParams! -v "%%~fi:/!DXCLIENT!/!VOLUME_DIR!/%%~nxi":Z
+			SET MOUNTED_VOL=!MOUNTED_VOL! "%%~fi"
         ) ELSE (
             CALL SET "newCmd=!newCmd! %%i"
         )
@@ -88,13 +121,24 @@ CALL SET searchVal=\
 	)
 )
 
+CALL SET environmentVars=!environmentVars! -e DXCLIENT_RUNTIME=!CONTAINER_RUNTIME! -e HOST_PATH=!hostPath!
 CALL !CONTAINER_RUNTIME! run !INTERACTIVE! !environmentVars! !volumeParams! --network="host" --name !DXCLIENT! --rm !IMAGE_NAME!:!IMAGE_TAG! ./bin/dxclient !newCmd!
 
 :: Cleanup the files created in local volume/mount point if the command is livesync
-IF %1 == livesync (
-	for %%i in (!execCmd!) do (
+IF %1 == livesync set result=true
+IF %2 == export-assets set result=true
+IF %result% == true (
+	for %%i in (!MOUNTED_VOL!) do (
 		IF EXIST %%~si\NUL (
-			rmdir /s /q "%cd%\!VOLUME_DIR!\%%~nxi"
+			SET ABS_PATH=%%~fi
+			for %%A in (!ABS_PATH!) do (
+				SET ORIGIN="%%~dpA"
+			)
+			IF NOT !ORIGIN! == "%cd%\!VOLUME_DIR!\" (
+				IF EXIST "%cd%\!VOLUME_DIR!\%%~nxi" (
+					rmdir /s /q "%cd%\!VOLUME_DIR!\%%~nxi"
+				)
+			)
 		) ELSE IF EXIST %%i (
 			del "%cd%\!VOLUME_DIR!\%%~nxi"
 		)
